Feature - Show And Edit Users
Update Folder Structure
Add Working Validation For Users
Add A Working Phone Number Validation

+++ routes/web.php
@@ -1,53 +1,67 @@
 <?php
 
-use App\Http\Controllers\Api\UserController;
-use App\Http\Controllers\DashboardController;
-use App\Http\Controllers\ProfileController;
-use App\Http\Controllers\TaskController;
+use App\Http\Controllers\Auth\DashboardController;
+use App\Http\Controllers\Auth\ProfileController;
+use App\Http\Controllers\Auth\TaskController;
+use App\Http\Controllers\Auth\UserController;
+use App\Http\Controllers\IndexController;
 use Illuminate\Foundation\Application;
+use Illuminate\Foundation\Http\Middleware\HandlePrecognitiveRequests;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
 Route::get('/', function () {
-    return Inertia::render('Welcome', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
+    Route::get('/', [IndexController::class, 'index']);
 })->name('home');
 
 Route::middleware('auth')->group(function () {
     Route::prefix('/dashboard')->name('dashboard.')->group(function () {
         Route::get('/', [DashboardController::class, 'index'])->middleware(['auth', 'verified'])->name('index');
+    });
 
-        Route::prefix('/api')->name('api.')->group(function () {
-            Route::get('/', [\App\Http\Controllers\Api\DashboardController::class, 'index'])->name('index');
+    Route::prefix('/api')->name('api.')->group(function () {
+        Route::prefix('/dashboard')->name('dashboard.')->group(function () {
+            Route::get('/', [\App\Http\Controllers\Api\Auth\DashboardController::class, 'index'])->name('index');
+        });
+
+
+        Route::prefix('/tasks')->name('tasks.')->group(function () {
+            Route::get('/', [\App\Http\Controllers\Api\Auth\TaskController::class, 'index'])->name('index');
+            Route::post('/', [\App\Http\Controllers\Api\Auth\TaskController::class, 'store'])->name('store')->middleware([HandlePrecognitiveRequests::class]);
+            Route::patch('/{task}', [\App\Http\Controllers\Api\Auth\TaskController::class, 'update'])->name('update')->middleware([HandlePrecognitiveRequests::class]);
+            Route::patch('/{task}/status', [\App\Http\Controllers\Api\Auth\TaskController::class, 'changeStatus'])->name('status');
+            Route::patch('/{task}/user', [\App\Http\Controllers\Api\Auth\TaskController::class, 'changeUser'])->name('user')->middleware([HandlePrecognitiveRequests::class]);
+            Route::delete('/{task}', [\App\Http\Controllers\Api\Auth\TaskController::class, 'destroy'])->name('destroy');
+        });
+
+        Route::prefix('/titles')->name('titles.')->group(function () {
+            Route::get('/', [\App\Http\Controllers\Api\Auth\UserController::class, 'getTitlesList'])->name('list');
+        });
+
+        Route::prefix('/users')->name('users.')->group(function () {
+            Route::get('/', [\App\Http\Controllers\Api\Auth\UserController::class, 'index'])->name('index');
+            Route::get('/list', [\App\Http\Controllers\Api\Auth\UserController::class, 'getUsersList'])->name('list');
+            Route::post('/', [\App\Http\Controllers\Api\Auth\UserController::class, 'store'])->name('store')->middleware([HandlePrecognitiveRequests::class]);
+            Route::patch('/{user}', [\App\Http\Controllers\Api\Auth\UserController::class, 'update'])->name('update')->middleware([HandlePrecognitiveRequests::class]);
+            Route::delete('/{user}', [\App\Http\Controllers\Api\Auth\UserController::class, 'destroy'])->name('destroy');
         });
     });
 
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    /* This section will be updated with less options (no delete for a start) and most functions transferred to the users section instead */
+    Route::prefix('/profile')->name('profile.')->group(function () {
+        Route::get('/', [ProfileController::class, 'edit'])->name('edit');
+        Route::patch('/', [ProfileController::class, 'update'])->name('update');
+        Route::group(['middleware' => ['permission:user-delete']], function () {
+            Route::delete('/', [ProfileController::class, 'destroy'])->name('destroy');
+        });
+    });
 
     Route::prefix('/tasks')->name('tasks.')->group(function () {
         Route::get('/', [TaskController::class, 'index'])->name('index');
-
-        Route::prefix('/api')->name('api.')->group(function () {
-            Route::get('/', [\App\Http\Controllers\Api\TaskController::class, 'index'])->name('index');
-            Route::post('/', [\App\Http\Controllers\Api\TaskController::class, 'store'])->name('store');
-            Route::patch('/{task}', [\App\Http\Controllers\Api\TaskController::class, 'update'])->name('update');
-            Route::patch('/{task}/status', [\App\Http\Controllers\Api\TaskController::class, 'changeStatus'])->name('status');
-            Route::patch('/{task}/user', [\App\Http\Controllers\Api\TaskController::class, 'changeUser'])->name('user');
-            Route::delete('/{task}', [\App\Http\Controllers\Api\TaskController::class, 'destroy'])->name('destroy');
-        });
     });
 
     Route::prefix('/users')->name('users.')->group(function () {
-
-        Route::prefix('api')->name('api.')->group(function () {
-            Route::get('/', [UserController::class, 'index'])->name('index');
-        });
+        Route::get('/', [UserController::class, 'index'])->name('index');
     });
 });
